WIP time interval algo
