+++ plugin_qgis_lpo/processing/summary_table_per_species_lpoaura.py
@@ -217,7 +217,8 @@ class SummaryTablePerSpecies(BaseProcessingAlgorithm):
         # Construct the sql array containing the study area's features geometry
         array_polygons = construct_sql_array_polygons(study_area)
         # Define the "where" clause of the SQL query, aiming to retrieve the output PostGIS layer = summary table
-        where = f"""is_valid and is_present and ST_intersects(obs.geom, ST_union({array_polygons}))"""
+        where = f"""is_valid and is_present
+        and ST_intersects(obs.geom, ST_union({array_polygons}))"""
         # Define a dictionnary with the aggregated taxons filters and complete the "where" clause thanks to it
         taxons_filters = {
             "groupe_taxo": groupe_taxo,
